requirement.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-tinymce==3.4.0
django-versatileimagefield==2.2
Pillow==9.0.1
python-magic==0.4.25
sqlparse==0.4.2
